+++ features/step_definitions/generic_steps.rb
@@ -60,3 +60,8 @@ end
 When(/^I (.*) the popup request modal$/) do |action|
   PageManager.current_page.modal_action(action)
 end
+
+
+When(/^I access the data for product id (.*)$/) do |product_id|
+  product_data_source.find_product(product_id)
+end